% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setAlgorithm.R
\docType{methods}
\name{algorithm}
\alias{algorithm}
\alias{algorithm,geoknife-method}
\title{get processing algorithm}
\usage{
algorithm(.Object)

\S4method{algorithm}{geoknife}(.Object)
}
\arguments{
\item{.Object}{a \code{geoknife} object.}
}
\value{
An \code{geoknife} object's algorithm
}
\description{
get processing algorithm
}
\details{
method for getting the process algorithm of the \code{geoknife} object.
}
\examples{
gk <- geoknife() # create geoknife object
algorithm <- list("Area Grid Statistics (weighted)"=
"gov.usgs.cida.gdp.wps.algorithm.FeatureWeightedGridStatisticsAlgorithm")
algorithm(gk) <- algorithm
algorithm(gk)
}
\author{
Jordan S. Read
}
\keyword{methods}

