% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setAlgorithm.R
\docType{methods}
\name{algorithm<-}
\alias{algorithm<-}
\alias{algorithm<-,geoknife-method}
\title{set processing algorithm}
\usage{
algorithm(.Object) <- value

\S4method{algorithm}{geoknife}(.Object) <- value
}
\arguments{
\item{.Object}{a \code{geoknife} object.}

\item{value}{a list for a valid algorithm, including values for name & location}
}
\value{
An \code{geoknife} object.
}
\description{
set processing algorithm
}
\details{
method for setting the process algorithm of the \code{geoknife} object.
}
\examples{
gk <- geoknife() # create geoknife object
algorithm <- list("Area Grid Statistics (weighted)"=
"gov.usgs.cida.gdp.wps.algorithm.FeatureWeightedGridStatisticsAlgorithm")
algorithm(gk) <- algorithm
}
\author{
Jordan S. Read
}
\keyword{methods}

