% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-webprocess.R
\docType{methods}
\name{algorithm}
\alias{algorithm}
\alias{algorithm<-}
\alias{algorithm,webprocess-method}
\alias{algorithm<-,webprocess-method}
\alias{algorithm,xml_document-method}
\title{the algorithm of a webprocess object}
\usage{
algorithm(.Object)

algorithm(.Object) <- value

\S4method{algorithm}{webprocess}(.Object)

\S4method{algorithm}{webprocess}(.Object) <- value

\S4method{algorithm}{xml_document}(.Object)
}
\arguments{
\item{.Object}{a \linkS4class{webprocess} object}

\item{value}{a list with name of algorithm and relative url endpoint}
}
\description{
Functions to get or set the algorithm of a \linkS4class{webprocess} object. 
The algorithm is the type of process that will be used, and can be accessed 
or modified using the algorithm method.
}
\examples{
wp <- webprocess()
algorithm(wp)
}
