% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAlgorithms.R
\docType{methods}
\name{algorithms}
\alias{algorithms}
\title{Get processing algorithms}
\usage{
algorithms(.Object)
}
\arguments{
\item{.Object}{a \code{geoknife} object with a valid WPS url.}
}
\value{
list of available algorithms for the \code{geoknife} WPS url.
}
\description{
Get algorithms for \code{geoknife}
}
\details{
algorithms is a method for finding algorithm names and locations for
a \code{geoknife} object.
}
\examples{
wp <- webprocess() # create geoknife object
algorithms(wp)
algorithms() # list Geo Data Portal processing algorithms
}
\author{
Jordan S. Read
}
\seealso{
\code{setAlgorithm<-}
}
\keyword{methods}

