% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check.R
\docType{methods}
\name{check}
\alias{check}
\alias{check,geojob}
\alias{check,geojob-method}
\title{Check status of processing request}
\usage{
check(.Object)

\S4method{check}{geojob}(.Object)
}
\arguments{
\item{.Object}{a \code{geoknife} object with an active GDP process request.}
}
\value{
\code{process}, a list containing
\code{status} and \code{URL}.
}
\description{
Check status of processing request
}
\details{
checkProcess is a method for checking the process status of an active (executed)
\code{geoknife} object. The method returns \code{process}, which is a list containing
two fields: \code{status} and \code{URL}. If the \code{geoknife} object has not been executed
(see \code{\link{start}}), this method returns \code{status}='none' and \code{URL}=NULL.
}
\examples{
gj <- geojob() # create geojob object
check(gj) # no process for empty geojob object
}
\author{
Jordan S. Read
}
\seealso{
\code{\link{start}}
}
\keyword{checkProcess}

