% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07-datagroup-obj.R
\docType{class}
\name{datagroup-class}
\alias{datagroup-class}
\title{datagroup class}
\description{
datagroup class
}
\section{Slots}{

\describe{
\item{\code{group}}{a list of webdata compatable elements}
}}

