% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeProcessInputs.R
\name{defaultProcessInputs}
\alias{defaultProcessInputs}
\title{Default Process Inputs}
\usage{
defaultProcessInputs(algorithm, wps_url, wps_version)
}
\arguments{
\item{algorithm}{the WPS algorithm to get process inputs for}

\item{wps_url}{the service base URL for the WPS}

\item{wps_version}{the service version to use}
}
\value{
list of default, optional, and required process inputs for use in the webprocess object.
}
\description{
parses DescribeProcess request
}
