% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download-geojob.R
\docType{methods}
\name{download}
\alias{download}
\alias{download,geojob,character-method}
\alias{download,geojob,missing-method}
\title{download output from geojob}
\usage{
download(.Object, destination, ...)

\S4method{download}{geojob,missing}(.Object, destination, ...)

\S4method{download}{geojob,character}(.Object, destination, ...)
}
\arguments{
\item{.Object}{a \code{\link{geojob}} that has completed}

\item{destination}{a file destination. If missing, a temp directory will be used}

\item{...}{additional arguments passed to \code{\link[httr]{write_disk}}, such as overwrite = TRUE}
}
\value{
the file handle
}
\description{
download output from geojob
}
\author{
Jordan S Read
}

