% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/email-geojob.R
\docType{methods}
\name{email}
\alias{email}
\alias{email,geojob,character,webprocess-method}
\alias{email,geojob-method}
\title{email user when processing job is complete}
\usage{
email(geojob, email, knife)

\S4method{email}{geojob,character,webprocess}(geojob, email,
  knife = webprocess())
}
\arguments{
\item{email}{a character for email address to use}
}
\description{
email user when processing job is complete
}
\keyword{internal}

