% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-webgeom.R
\name{fetchGML_IDs}
\alias{fetchGML_IDs}
\title{fetch GML_IDs from WFS}
\usage{
fetchGML_IDs(.Object)
}
\arguments{
\item{.Object}{a webgeom object}
}
\description{
fetch GML_IDs from WFS when geom, attribute, and values are specified
}
\keyword{internal}

