% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-geoknife.R
\name{gcontent}
\alias{gcontent}
\title{xml2 version of gcontent}
\usage{
gcontent(response)
}
\arguments{
\item{response}{the result of httr::GET(url)}
}
\description{
xml2 version of gcontent
}
\keyword{internal}
