% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoknifeUtils.R
\name{gcontent_xml2}
\alias{gcontent_xml2}
\title{xml2 version of gcontent}
\usage{
gcontent_xml2(response)
}
\arguments{
\item{response}{the result of httr::GET(url)}
}
\description{
xml2 version of gcontent
}
\keyword{internal}
