% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03-geojob-obj.R
\docType{class}
\name{geojob-class}
\alias{geojob-class}
\title{geojob class}
\description{
geojob class
}
\section{Slots}{

\describe{
\item{\code{url}}{URL of web processing endpoint}

\item{\code{xml}}{XML character for post}

\item{\code{id}}{job identifier}

\item{\code{package.version}}{the version of the geoknife package}

\item{\code{algorithm.version}}{the version of the algorithm used for processing}
}}

