% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/03-geojob-obj.R
\docType{methods}
\name{geojob}
\alias{geojob}
\alias{geojob,ANY-method}
\alias{geojob,geojob-method}
\title{create geojob object}
\usage{
geojob(...)

\S4method{geojob}{ANY}(...)
}
\value{
the geojob object
}
\description{
A class representing a geoknife job (\code{geojob}).
}
\author{
Jordan S Read
}

