\name{geoknife}
\alias{geoknife}
\title{geoknife class}
\description{
  Some details about the \code{geoknife} class \describe{
  \item{WFS_URL}{endpoint for web feature service (WFS)}
  \item{WPS_URL}{endpoint for web processing service (WPS)}
  \item{algorithm}{a length one list (name and location) of
  a WPS algorithm} \item{processInputs}{a list of process
  parameters} \item{feature}{a list of elements in the
  feature collection} }
}

