% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoknife-generic.R
\docType{methods}
\name{geoknife}
\alias{geoknife}
\title{geoknife}
\usage{
geoknife(stencil, fabric, ..., knife = webprocess(...),
  emailComplete = FALSE)
}
\arguments{
\item{stencil}{a \linkS4class{webgeom}, \linkS4class{simplegeom}, or any type
that can be coerced into \linkS4class{simplegeom}.}

\item{fabric}{a dataset. A \linkS4class{webdata} or any type that
can be coerced into \linkS4class{webdata}}

\item{...}{additional arguments passed to \code{new} \linkS4class{webdata}.
Can also be used to modify the \code{knife} argument, if that is supplied.}

\item{knife}{(optional) a \code{\link{webprocess}} object}

\item{emailComplete}{FALSE by default. \code{character} of valid email to
send email for failed or completed process. NOT IMPLEMENTED}
}
\description{
geoknife
}
\examples{
wp <- quick_wp()
job <- geoknife(stencil = c(-89,42), fabric = 'prism', knife = wp)
check(job)
}

