% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-webgeom.R
\docType{methods}
\name{geom<-}
\alias{geom}
\alias{geom,webgeom-method}
\alias{geom<-}
\alias{geom<-,webgeom-method}
\title{the geom of an object}
\usage{
geom(.Object) <- value

geom(.Object)

\S4method{geom}{webgeom}(.Object) <- value

\S4method{geom}{webgeom}(.Object)
}
\arguments{
\item{.Object}{a \linkS4class{webgeom} object}

\item{value}{a geom}
}
\description{
The "feature" of a webgeom. This is the key mapping to the web resource 
that is used as the spatial feature of refernce. Other details specified in 
\code{\link{attribute}} and \code{\link{values}}.
}
\seealso{
\code{\link{attribute}} and \code{\link{values}}
}

