% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDataIDs.R
\docType{methods}
\name{getDataIDs}
\alias{getDataIDs}
\title{Find variables from dataset}
\usage{
getDataIDs(.Object, cachedResponse)
}
\arguments{
\item{.Object}{a \code{geoknife} object with a valid dataset URI.}

\item{cachedResponse}{an optional input to allow cached response. Default as FALSE}
}
\value{
list of dataset IDs for the \code{geoknife} dataset URI.
}
\description{
Find variables from \code{geoknife} dataset
}
\details{
getDataIDs is a method for finding dataset variables from a
\code{geoknife} object with a valid dataset URI. Use setProcessInputs to set a dataset URI.
}
\examples{
gk <- geoknife() # create geoknife object
algorithm <- list("Area Grid Statistics (weighted)"=
"gov.usgs.cida.gdp.wps.algorithm.FeatureWeightedGridStatisticsAlgorithm")
setAlgorithm(gk) <- algorithm
setProcessInputs(gk) <- list("DATASET_URI" = "dods://hydro1.sci.gsfc.nasa.gov:80/dods/NLDAS_FORA0125_H.002")
getDataIDs(gk)
}
\author{
Jordan S. Read
}
\seealso{
\code{setProcessInputs}
}
\keyword{methods}

