% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setJobState.R
\name{getJobState}
\alias{getJobState}
\title{get job state}
\usage{
getJobState()
}
\value{
\code{jobState} package variable
}
\description{
convienence function for retrieving package-level \code{jobState}
}
\keyword{internal}
