\docType{methods}
\name{getShapefiles}
\alias{getShapefiles}
\title{get shapefiles from a web location}
\arguments{
  \item{\code{rGDP}}{object with a valid WFS url.}
}
\value{
  list of shapefiles for the \code{rGDP} WFS url.
}
\description{
  get shapefiles from a web location
}
\details{
  a \code{rGDP} method for finding shapefile names at a
  valid WFS endpoint.
}
\keyword{getShapefiles}

