% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IDs-webgeom.R
\docType{methods}
\name{IDs<-}
\alias{IDs}
\alias{IDs,webgeom-method}
\alias{IDs<-}
\alias{IDs<-,webgeom-method}
\title{the IDs of an object}
\usage{
IDs(.Object) <- value

IDs(.Object)

\S4method{IDs}{webgeom}(.Object) <- value

\S4method{IDs}{webgeom}(.Object)
}
\arguments{
\item{.Object}{a \linkS4class{webgeom} object}

\item{value}{a IDs}
}
\description{
the IDs of an object
}

