% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputs-webprocess.R
\docType{methods}
\name{inputs<-}
\alias{inputs}
\alias{inputs,webprocess-method}
\alias{inputs<-}
\alias{inputs<-,webprocess-method}
\title{inputs of webprocess}
\usage{
inputs(.Object, ...) <- value

inputs(.Object, ...)

\S4method{inputs}{webprocess}(.Object, ...)

\S4method{inputs}{webprocess}(.Object, ...) <- value
}
\description{
access or set the inputs of a webprocess
}
\keyword{internal}

