% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadOutput.R
\docType{methods}
\name{loadOutput}
\alias{loadOutput}
\title{parse process output into R environment}
\usage{
loadOutput(.Object)
}
\arguments{
\item{.Object}{a \code{\link{geoknife}} object with a successful processID.
(See \code{\link{checkProcess}}).}
}
\value{
list of timesries values.
}
\description{
parse process output into R environment
}
\details{
a \code{geoknife} method for finding value names for a given shapefile at a valid WFS endpoint.
}
\author{
Jordan S. Read
}
\keyword{methods}

