% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCategorical.R
\name{parseCategorical}
\alias{parseCategorical}
\title{parse categorical coverage file into R environment}
\usage{
parseCategorical(file, delim)
}
\arguments{
\item{file}{a \linkS4class{geojob} categorical processing result file location
(See \code{\link{check}}).}

\item{delim}{the file delimiter}
}
\value{
a data.frame of categorical fraction (and/or count) values.
}
\description{
a function for loading data into R from a file (or URL) from a completed processing request
}
\examples{
local.file <- system.file('extdata','csv_categorical_multifeature.csv', package = 'geoknife')
output <- parseCategorical(local.file, delim = ',')
}
\keyword{methods}

