% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTimeseries.R
\name{parseTimeseries}
\alias{parseTimeseries}
\title{parse timeseries file into R environment}
\usage{
parseTimeseries(file, delim, keep.units = FALSE)
}
\arguments{
\item{file}{a \linkS4class{geojob} timeseries processing result file location
(See \code{\link{check}}).}

\item{delim}{the file delimiter}

\item{keep.units}{boolean for including a units column in returned data.frame}
}
\value{
a data.frame of timeseries values.
}
\description{
parse timeseries file into R environment
}
\details{
a function for loading data into R from a file (or URL) from a completed processing request
}
\examples{
local_file <- system.file('extdata','tsv_linear_ring.tsv', package = 'geoknife')
output <- parseTimeseries(local_file, delim = '\\t')
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}

