% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printProcessInputs.R
\name{printProcessInputs}
\alias{printProcessInputs}
\title{Print out the process request xml for diagnostic purposes}
\usage{
printProcessInputs(.Object)
}
\arguments{
\item{.Object}{a \code{geoknife} object.}
}
\description{
function for printing processXML \code{geoknife} object.
}
\examples{
gk <- geoknife() # create geoknife object
linearRing = bufferPoint(c(-111.48,36.95))
setFeature(gk) <- list(LinearRing=linearRing)

algorithm <- list("Area Grid Statistics (weighted)"=
"gov.usgs.cida.gdp.wps.algorithm.FeatureWeightedGridStatisticsAlgorithm")
algorithm(gk) <- algorithm
setProcessInputs(gk) <- list("DATASET_URI" = "dods://hydro1.sci.gsfc.nasa.gov:80/dods/NLDAS_FORA0125_H.002")
printProcessInputs(gk)
}
\keyword{methods}

