% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query-webdata.R
\docType{methods}
\name{query,webdata,character-method}
\alias{query,webdata,character-method}
\alias{query,webdata-method}
\title{query webdata for various fields}
\usage{
\S4method{query}{webdata,character}(.Object, field, ...)
}
\arguments{
\item{.Object}{a \code{webdata} object.}

\item{field}{a plural parameter name for fields in webdata (e.g., 'variables', 'times')}

\item{...}{additional arguments passed to methods, e.g., a \code{webdprocess} object}
}
\value{
a character vector of values corresponding to the query field specified
}
\description{
query webdata for various fields
}
\details{
a \code{webdata} method for finding shapefile names at a valid WFS endpoint.
}
\examples{
fabric <- webdata('prism')
query(fabric, 'variables')
}
\keyword{methods}

