% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query-webgeom.R
\name{query}
\alias{query}
\title{query webgeom for various fields}
\usage{
query(.Object, field, ...)
}
\arguments{
\item{.Object}{a \code{webgeom} object.}

\item{field}{a plural parameter name for fields in webgeom (e.g., 'geoms', 'attributes','values')}

\item{...}{additional arguments passed to methods (e.g., rm.duplicates = TRUE)}
}
\value{
a character vector of values corresponding to the query field specified
}
\description{
query webgeom for various fields
}
\details{
a \code{webgeom} method for finding shapefile names at a valid WFS endpoint.
}
\examples{
wg <- webgeom()
query(wg, 'geoms')
geom(wg) <- "derivative:CONUS_States"
query(wg, 'attributes')
attribute(wg) <- 'STATE'
query(wg, 'values', rm.duplicates = TRUE)
}
\author{
Jordan S. Read
}
\keyword{methods}

