% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/04-webprocess-obj.R
\name{quick_wp}
\alias{quick_wp}
\title{quick webprocess object}
\usage{
quick_wp(...)
}
\arguments{
\item{...}{additional arguments passed to \code{webprocess}}
}
\description{
quick webprocess object
}
\keyword{internal}

