\name{rGDP}
\alias{rGDP}
\title{rGDP class}
\description{
  Some details about the \code{rGDP} class \describe{
  \item{WFS_URL}{endpoint for web feature service (WFS)}
  \item{PROCESS_URL}{endpoint for web processing service
  (WPS)} \item{algorithm}{acronym for WPS algorithm}
  \item{datasetURI}{processing dataset URI}
  \item{PostInputs}{a list of process parameters}
  \item{feature}{a list of elements in the feature
  collection} }
}

