\docType{package}
\name{rGDP-package}
\alias{rGDP-package}
\title{rGDP package}
\description{
  \tabular{ll}{ Package: \tab rGDP\cr Type: \tab Package\cr
  Version: \tab 0.0.1\cr Date: \tab 2013-05-23\cr License:
  \tab Unlimited for this package, dependencies have more
  restrictive licensing.\cr Copyright: \tab This software
  is in the public domain because it contains materials
  that originally came from the United States Geological
  Survey, an agency of the United States Department of
  Interior. For more information, see the official USGS
  copyright policy at
  http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
  LazyLoad: \tab yes\cr }
}
\details{
  Collection of functions to do GDP calls
}
\author{
  Jordan Read \email{jread@usgs.gov}
}

