% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setFeature.R
\docType{methods}
\name{setFeature<-}
\alias{setFeature<-}
\title{Set feature geometry for processing}
\usage{
setFeature(.Object) <- value
}
\arguments{
\item{.Object}{a \code{geoknife} object.}

\item{value}{a list containing a valid feature collection, or a list of a subset of a valid feature collection.}
}
\value{
An \code{geoknife} object.
}
\description{
Set feature for \code{geoknife}
}
\details{
setFeature is a method for setting or modifying the feature elements
of for \code{geoknife} object.
}
\examples{
gk <- geoknife() # create geoknife object
linearRing = bufferPoint(c(-111.48,36.95))
setFeature(gk) <- list(LinearRing=linearRing)

gk2 <- geoknife() # create geoknife object
feature_collection  <-  'sb:WBIC_190900'
attribute  <-  'WBDY_WBIC'
setFeature(gk2) <- list('FEATURE_COLLECTION'=feature_collection,'ATTRIBUTE'=attribute)
}
\author{
Jordan S. Read
}
\seealso{
\code{setWFS}
}
\keyword{setFeature}

