\docType{methods}
\name{setFeature}
\alias{setFeature}
\title{setFeature}
\arguments{
  \item{An}{\code{rGDP} object.}

  \item{a}{list containing a valid feature collection, or a
  list of a subset of a valid feature collection.}
}
\value{
  An \code{rGDP} object.
}
\description{
  function for rGDP
}
\keyword{setFeature}

