% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setJobState.R
\name{setJobState}
\alias{setJobState}
\title{set job state at package level for outgoing requests}
\usage{
setJobState(state = NULL)
}
\arguments{
\item{state}{the state (as a character) to set \code{jobState} to.}
}
\value{
invisible return of \code{jobState} package variable
}
\description{
this function keeps the package from rifling off multiple requests
while another is still pending. Package variable \code{jobState} is set with
direct calls to this private function, or as part of any \code{status} type
method for geojob
}
\keyword{internal}

