\docType{methods}
\name{setPostInputs}
\alias{setPostInputs}
\title{setPostInputs}
\arguments{
  \item{An}{\code{rGDP} object.}

  \item{a}{list of valid postInputs.}
}
\value{
  An \code{rGDP} object.
}
\description{
  method for setting the (non-feature related) post inputs
  of the \code{rGDP} object.
}
\keyword{setPostInputs}

