\docType{methods}
\name{setWFS}
\alias{setWFS}
\title{set web feature service location}
\arguments{
  \item{\code{geoknife}}{object.}

  \item{a}{Web Feature Service (WFS) endpoint.}
}
\value{
  An \code{geoknife} object.
}
\description{
  set web feature service location
}
\details{
  method for setting the web feature service (WFS) endpoint
  for a \code{geoknife} object.
}
\keyword{setWFS}

