% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setWPS<-}
\alias{setWPS<-}
\title{set web processing service location}
\usage{
setWPS(.Object) <- value
}
\arguments{
\item{.Object}{a \code{geoknife} object.}

\item{value}{a Web Processing Service (WPS) endpoint.}
}
\value{
An \code{geoknife} object.
}
\description{
set web processing service location
}
\details{
method for setting the web processing service (WPS) endpoint for a \code{geoknife} object.
}
\examples{
gk <- geoknife() # create geoknife object
WPS <- 'http://cida-eros-gdp2qa.er.usgs.gov:8080/gdp-utility-wps/WebProcessingService'
setWPS(gk) <- WPS
}
\keyword{setWPS}

