\docType{methods}
\name{setWPS}
\alias{setWPS}
\title{setWPS}
\arguments{
  \item{\code{rGDP}}{object.}

  \item{a}{Web Processing Service (WPS) endpoint.}
}
\value{
  An \code{rGDP} object.
}
\description{
  method for setting the web processing service (WPS)
  endpoint for a \code{rGDP} object.
}
\keyword{setWPS}

