% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/06-simplegeom-obj.R
\docType{class}
\name{simplegeom-class}
\alias{simplegeom-class}
\title{simplegeom class}
\description{
simplegeom class
}
\section{Slots}{

\describe{
\item{\code{DRAW_NAMESPACE}}{}

\item{\code{DRAW_SCHEMA}}{}

\item{\code{sp}}{a list for algorithm used}
}}

