% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/start.R
\docType{methods}
\name{start}
\alias{start}
\title{Submit a GDP web processing request}
\usage{
start(.Object)
}
\arguments{
\item{.Object}{a \code{geoknife} object ot be used to formulate GDP process request.}
}
\value{
An \code{geoknife} object with an active GDP process request.
}
\description{
Start process for \code{geojob}
}
\details{
startProcess is a method for submitting a web processing request for the \code{geoknife} object.
}
\examples{
wp <- quick_wp()
wd <- quick_wd()
wg <- quick_wg()
gj <- geojob()
xml(gj) <- XML(wg, wd, wp)
url(gj) <- wp@wps_url
gj <- start(gj)
id(gj) # get the processing ID for the geojob
# -- is execute true (geoknife) or "start(geojob)"
}
\seealso{
\code{check}
}
\keyword{startProcess}

