% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/successful.R
\docType{methods}
\name{successful}
\alias{error}
\alias{error,geojob-method}
\alias{running}
\alias{running,geojob-method}
\alias{successful}
\alias{successful,geojob-method}
\title{Convenience function for GDP process state}
\usage{
successful(.Object)
error(.Object)
running(.Object)

\S4method{successful}{geojob}(.Object)

\S4method{running}{geojob}(.Object)

\S4method{error}{geojob}(.Object)
}
\arguments{
\item{.Object}{a \code{\link{geojob}} object}
}
\value{
TRUE/FALSE indicating if process is in the given state (error, processing, successful)
}
\description{
Simple wrapper to check process status
}
\examples{
\dontrun{
wp <- quick_wp(url = 'http://cida-test.er.usgs.gov/gdp/process/WebProcessingService')
wd <- quick_wd()
wg <- quick_wg()
gj <- geojob()
xml(gj) <- XML(wg, wd, wp)
url(gj) <- url(wp)
gj <- start(gj)

running(gk)
error(gk)
successful(gk)
}
}
\author{
Luke Winslow
}
\seealso{
\code{\link{check}}
}

