% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query-webdata-times.R
\docType{methods}
\name{times_query}
\alias{times-query,webdata-method}
\alias{times_query}
\alias{times_query,webdata,missing-method}
\alias{times_query,webdata,webprocess-method}
\title{times query}
\usage{
times_query(fabric, knife)

\S4method{times_query}{webdata,missing}(fabric, knife)

\S4method{times_query}{webdata,webprocess}(fabric, knife)
}
\arguments{
\item{times}{}
}
\description{
times query
}
\keyword{internal}

