% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times-webdata.R
\docType{methods}
\name{times}
\alias{times}
\alias{times<-}
\alias{times<-}
\alias{times}
\alias{times<-,webdata-method}
\alias{times}
\alias{times,webdata-method}
\alias{times}
\title{the times of an webdata object}
\usage{
times(.Object)

times(.Object) <- value

\S4method{times}{webdata}(.Object) <- value

\S4method{times}{webdata}(.Object)
}
\arguments{
\item{.Object}{a \code{\link{webdata}} object}

\item{value}{a POSIXct vector}
}
\description{
Functions to get or set the times of a \code{\link{webdata}} object
}
\examples{
wd <- webdata('prism')
times(wd) <- as.POSIXct(c("2012-11-04", "2012-11-12"))
times(wd)[1] <- as.POSIXct("2012-11-04")
times(wd)
}
