% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/url-method.R
\name{url}
\alias{url}
\alias{url<-}
\title{the url of an object}
\usage{
url(.Object)
url(.Object) <- 'dods://cida.usgs.gov/thredds/dodsC/prism'
}
\arguments{
\item{.Object}{a webgeom, webdata, geojob, or webprocess object}
}
\description{
the url of an object
}

