% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-webgeom.R
\docType{methods}
\name{values<-}
\alias{values}
\alias{values,webgeom-method}
\alias{values<-}
\alias{values<-,webgeom-method}
\title{the values of an object}
\usage{
values(.Object) <- value

values(.Object)

\S4method{values}{webgeom}(.Object) <- value

\S4method{values}{webgeom}(.Object)
}
\arguments{
\item{.Object}{a \linkS4class{webgeom} object}

\item{value}{a values}
}

