% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-webdata-variables.R
\docType{methods}
\name{variables_query}
\alias{variables_query}
\alias{variables_query,webdata-method}
\alias{variables_query,webdata,missing-method}
\alias{variables_query,webdata,webprocess-method}
\title{variables query}
\usage{
variables_query(fabric, knife)

\S4method{variables_query}{webdata,missing}(fabric, knife)

\S4method{variables_query}{webdata,webprocess}(fabric, knife)
}
\arguments{
\item{variables}{}
}
\description{
find possible variables in a webdata dataset
}
\keyword{internal}
