% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-generic.R
\docType{methods}
\name{version<-}
\alias{version}
\alias{version,ANY-method}
\alias{version<-}
\alias{version<-,ANY-method}
\title{the version of an object}
\usage{
version(.Object) <- value

version(.Object)

\S4method{version}{ANY}(.Object) <- value

\S4method{version}{ANY}(.Object)
}
\arguments{
\item{.Object}{a \linkS4class{webgeom} or \linkS4class{webprocess} object}

\item{value}{a version}
}

