% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitUntilFinished.R
\name{wait}
\alias{wait}
\title{hold up R while GDP is processing}
\usage{
wait(.Object, sleep.time = gconfig("sleep.time"))
}
\arguments{
\item{.Object}{a geojob}

\item{sleep.time}{a number of seconds to wait in between checking the process}
}
\value{
invisible return of .Object, unaltered
}
\description{
keeps R in a loop while GDP works on the request. Checks \code{\link{running}}. 
Will drop out of loop whenever !running(geojob)
}
\examples{
\dontrun{
job <- geoknife(stencil = c(-89,42), fabric = 'prism')
2+2
wait(job)
check(job) # should be complete
}
}

