% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitUntilFinished.R
\name{waitUntilFinished}
\alias{waitUntilFinished}
\title{hold up R while GDP is processing}
\usage{
waitUntilFinished(.Object, sleep.time = gconfig("sleep.time"))
}
\description{
keeps R in a loop while GDP works on the request. 
Checks \code{\link{running}}. Will drop out of loop whenever !running(geojob)
}
\keyword{internal}

