% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-webdata-obj.R
\docType{class}
\name{webdata-class}
\alias{webdata-class}
\title{webdata class}
\description{
a class for specifying details of web datasets (webdata!). These datasets have to be  
accessible through the OPeNDAP protocol or as WCS (web coverage services).
}
\section{Slots}{

\describe{
\item{\code{times}}{vector of POSIXct dates (specifying start and end time of processing)}

\item{\code{url}}{URL of web data}

\item{\code{variables}}{variable(s) used for processing from dataset}
}}

