% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/01-webdata-obj.R
\docType{methods}
\name{webdata}
\alias{webdata}
\alias{webdata,missing-method}
\alias{webdata,webdata-method}
\title{create webdata object}
\usage{
webdata(value, ...)

\S4method{webdata}{missing}(value, ...)
}
\arguments{
\item{value}{character name of dataset or [others supported in the future]}

\item{...}{additional arguments passed initialize method (e.g., times vector)}
}
\value{
the webdata object representing a dataset and parameters
}
\description{
A class representing a web dataset.
}
\section{Slots}{

\describe{
\item{\code{times}}{value of type \code{"POSIXct"}, start and stop dates for data}

\item{\code{url}}{value of type \code{"character"}, the web location for the dataset}

\item{\code{variable}}{value of type \code{"character"}, the variable(s) for data}
}}
\author{
Jordan S Read
}

