% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/04-webprocess-obj.R
\docType{methods}
\name{webprocess}
\alias{algorithm,webprocess-method}
\alias{algorithm<-,webprocess-method}
\alias{webprocess}
\alias{webprocess,ANY-method}
\alias{webprocess,webprocess-method}
\title{create webprocess object}
\usage{
webprocess(...)

\S4method{webprocess}{ANY}(...)

\S4method{algorithm}{webprocess}(.Object)

\S4method{algorithm}{webprocess}(.Object) <- value
}
\value{
the webprocess object
}
\description{
A class representing geoknife web processing specifications
}
\author{
Jordan S Read
}

