% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/processInputsToXML.R
\docType{methods}
\name{XML}
\alias{XML}
\alias{XML,ANY,webdata,webprocess-method}
\alias{XML,webgeom-method}
\title{XML from set of objects}
\usage{
XML(stencil, fabric, knife)

\S4method{XML}{ANY,webdata,webprocess}(stencil, fabric, knife)
}
\arguments{
\item{stencil}{}

\item{fabric}{}

\item{webprocess}{}
}
\value{
XML as ?string?
}
\description{
XML from set of objects
}
\examples{
wp <- quick_wp()
wd <- quick_wd()
wg <- quick_wg()
XML(wg, wd, wp)
}

