% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-datagroup-obj.R
\docType{methods}
\name{abstract}
\alias{abstract}
\alias{abstract,datagroup-method}
\alias{title}
\alias{title,datagroup-method}
\title{get abstract from a datagroup}
\usage{
abstract(.Object)

\S4method{abstract}{datagroup}(.Object)

title(.Object)

\S4method{title}{datagroup}(.Object)
}
\arguments{
\item{.Object}{a datagroup object}
}
\description{
extracts the abstract information from a datagroup object
}

