% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07-datagroup-obj.R
\docType{methods}
\name{abstract}
\alias{abstract}
\alias{abstract,datagroup-method}
\title{get abstract from a datagroup}
\usage{
abstract(.Object)

\S4method{abstract}{datagroup}(.Object)
}
\description{
extracts the abstract information from a datagroup object
}

