% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/algorithm-webprocess.R
\docType{methods}
\name{algorithm}
\alias{algorithm}
\alias{algorithm,webprocess-method}
\alias{algorithm<-}
\alias{algorithm<-,webprocess-method}
\title{the algorithm of a webprocess object}
\usage{
algorithm(.Object) <- value
algorithm(.Object)

algorithm(.Object) <- value

\S4method{algorithm}{webprocess}(.Object)

\S4method{algorithm}{webprocess}(.Object) <- value
}
\arguments{
\item{.Object}{a \linkS4class{webprocess} object}

\item{value}{a list with name of algorithm and relative url endpoint}
}
\description{
Functions to get or set the algorithm of a \linkS4class{webprocess} object
}
\examples{
wp <- webprocess()
algorithm(wp)
}

