% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-webgeom.R
\docType{methods}
\name{attribute<-}
\alias{attribute<-}
\alias{attribute}
\alias{attribute<-,webgeom-method}
\alias{attribute,webgeom-method}
\title{the attribute of an webgeom object}
\usage{
attribute(.Object) <- value

attribute(.Object)

\S4method{attribute}{webgeom}(.Object) <- value

\S4method{attribute}{webgeom}(.Object)
}
\arguments{
\item{.Object}{a \linkS4class{webgeom} object}

\item{value}{a attribute}
}
\description{
get or set the attribute of a webgeom object.
}
