% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/attribute-webgeom.R
\docType{methods}
\name{attribute<-}
\alias{attribute}
\alias{attribute<-}
\alias{attribute<-,webgeom-method}
\title{the attribute of an object}
\usage{
attribute(.Object) <- value

attribute(.Object)

\S4method{attribute}{webgeom}(.Object) <- value
}
\arguments{
\item{.Object}{a \linkS4class{webgeom} object}

\item{value}{a attribute}
}
\description{
the attribute of an object
}

