% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setJobState.R
\name{canStart}
\alias{canStart}
\title{can start a new job?}
\usage{
canStart()
}
\value{
TRUE or FALSE
}
\description{
checks package job state and returns a boolean for whether a new job can be started
}
\keyword{internal}

