\docType{methods}
\name{checkProcess}
\alias{checkProcess}
\title{Checks status of processing request}
\arguments{
  \item{\code{rGDP}}{object with an active GDP process
  request.}
}
\value{
  process of \code{rGDP} process.
}
\description{
  checkProcess
}
\details{
  method for checking the process status of an active
  (executed) \code{rGDP} object.
}
\keyword{checkProcess}

