% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkProcess.R
\docType{methods}
\name{checkProcess}
\alias{checkProcess}
\title{Check status of processing request}
\usage{
checkProcess(.Object)
}
\arguments{
\item{.Object}{a \code{geoknife} object with an active GDP process request.}
}
\value{
\code{process}, a list containing
\code{status} and \code{URL}.
}
\description{
Check status of processing request
}
\details{
checkProcess is a method for checking the process status of an active (executed)
\code{geoknife} object. The method returns \code{process}, which is a list containing
two fields: \code{status} and \code{URL}. If the \code{geoknife} object has not been executed
(see \code{\link{startProcess}}), this method returns \code{status}='none' and \code{URL}=NULL.
}
\examples{
gk <- geoknife() # create geoknife object
checkProcess(gk) # no process for empty geoknife object
}
\author{
Jordan S. Read
}
\seealso{
\code{\link{startProcess}}
}
\keyword{checkProcess}

