% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-datagroup-obj.R
\docType{class}
\name{datagroup-class}
\alias{datagroup-class}
\title{datagroup class}
\description{
contains collections of webdata that can be processed with 
\code{\link{geoknife}}
}
\section{Slots}{

\describe{
\item{\code{group}}{a list of webdata compatible elements}
}}

