% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-datagroup-obj.R
\docType{methods}
\name{datagroup}
\alias{datagroup}
\alias{datagroup,ANY-method}
\alias{datagroup,datagroup-methods}
\alias{length,datagroup-method}
\alias{[,datagroup-method}
\alias{[[,datagroup,ANY,ANY-method}
\title{create datagroup object}
\usage{
datagroup(...)

\S4method{datagroup}{ANY}(...)

\S4method{length}{datagroup}(x)

\S4method{[}{datagroup}(x, i, j, ..., drop = TRUE)

\S4method{[[}{datagroup,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{...}{additional arguments passed to initialize method}

\item{x}{a datagroup object}

\item{i}{index specifying elements to extract or replace.}

\item{j}{not implemented}

\item{drop}{not implemented}
}
\value{
the datagroup object
}
\description{
A class representing a geoknife job (\code{datagroup}).
}
\author{
Jordan S Read
}
