% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-geojob.R
\docType{methods}
\name{download}
\alias{download}
\alias{download,geojob,missing-method}
\alias{download}
\alias{download,character,missing-method}
\alias{download}
\alias{download,geojob,character-method}
\alias{download}
\alias{download,character,character-method}
\alias{download}
\title{download output from geojob}
\usage{
download(.Object, destination, ...)

\S4method{download}{geojob,missing}(.Object, destination, ...)

\S4method{download}{character,missing}(.Object, destination, ...)

\S4method{download}{geojob,character}(.Object, destination, ...)

\S4method{download}{character,character}(.Object, destination, ...)
}
\arguments{
\item{.Object}{a \code{\link{geojob}} or job id that has completed}

\item{destination}{a file destination. If missing, a temp directory will be used}

\item{...}{additional arguments passed to \code{\link[httr]{write_disk}}, such as overwrite = TRUE}
}
\value{
the location of the downloaded file
}
\description{
download the result of a processing job to a local destination.
}
\author{
Jordan S Read
}
