% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email-geojob.R
\name{email}
\alias{email}
\alias{email,geojob-method}
\alias{email,geojob,webprocess-method}
\title{email user when processing job is complete}
\usage{
email(geojob, knife)

\S4method{email}{geojob,webprocess}(geojob, knife)
}
\arguments{
\item{email}{a character for email address to use}
}
\description{
configure job to send an email when complete (success or failure)
}
\keyword{internal}
