% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoknifeUtils.R
\name{gcontent}
\alias{gcontent}
\title{drop in replacement for httr switching to xml2 from XML}
\usage{
gcontent(response)
}
\arguments{
\item{response}{the result of httr::GET(url)}
}
\description{
reverts to old parsing pre v1.1.0 for httr
}
\keyword{internal}

