% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-geojob-obj.R
\docType{methods}
\name{geojob}
\alias{geojob}
\alias{geojob,missing-method}
\alias{geojob,geojob-method}
\alias{geojob,XMLDocument-method}
\alias{geojob,geojob-method}
\alias{geojob,character-method}
\alias{geojob,geojob-method}
\alias{xml<-}
\alias{xml<-,geojob-method}
\alias{xml}
\alias{xml,geojob-method}
\alias{id<-}
\alias{id,geojob-method}
\alias{id}
\alias{id,geojob-method}
\alias{id<-,geojob-method}
\alias{id,geojob-method}
\alias{id,geojob-method}
\alias{id,character-method}
\alias{id,geojob-method}
\title{create geojob object}
\usage{
geojob(xml, ...)

\S4method{geojob}{missing}(xml, ...)

\S4method{geojob}{XMLDocument}(xml, ...)

\S4method{geojob}{character}(xml, ...)

xml(.Object) <- value

xml(.Object)

id(.Object)
id(.Object) <- value

id(.Object)

\S4method{id}{geojob}(.Object) <- value

\S4method{id}{geojob}(.Object)

\S4method{id}{character}(.Object)
}
\arguments{
\item{xml}{location of xml (URL or local path)}

\item{...}{additional arguments passed to initialize method}

\item{.Object}{a \code{\link{geojob}} object}

\item{value}{a character string of xml}
}
\value{
the geojob object
}
\description{
A class representing a geoknife job (\code{geojob}).
}
\examples{
xml <- "<foo> <bar> text <baz/> </bar> </foo>"
gj <- geojob()
xml(gj) <- xml
xml(gj)
xml <- "<foo version=\\"1.0.0\\"> <bar> text <baz/> </bar> </foo>"
gj <- geojob(xml = xml)
xml(gj)
id(gj)
}
\author{
Jordan S Read
}
