% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/05-geoknife.R
\name{geoknife}
\alias{geoknife}
\title{geoknife-fcn}
\usage{
geoknife(stencil, fabric, knife, job, ...)
}
\arguments{
\item{stencil}{a geometric feature (the "cookie-cutter"). Many input types are supported. [list them]}

\item{fabric}{a dataset. Can be \code{\link{webdata}} or recognized shortname (e.g., 'prism')}

\item{knife}{(optional) a \code{\link{webprocessing}} object}

\item{job}{(optional) a \code{\link{geojob}} object}

\item{...}{additional arguments passed on to process.}
}
\description{
geoknife-fcn
}
\examples{
wp <- quick_wp(url = 'http://cida-test.er.usgs.gov/gdp/process/WebProcessingService')
wd <- quick_wd()
wg <- quick_wg()
geoknife(stencil = wg, fabric = wd, knife = wp)
sp <- quick_sp()
job <- geoknife(stencil = sp, fabric = wd, knife = wp)
check(job)
}

