% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoknife-generic.R
\docType{methods}
\name{geoknife}
\alias{geoknife}
\alias{geoknife,ANY,ANY,ANY,missing}
\alias{geoknife,ANY,ANY,ANY,missing-method}
\alias{geoknife,ANY,ANY,missing,missing}
\alias{geoknife,ANY,ANY,missing,missing-method}
\alias{geoknife,ANY,missing,missing,missing}
\alias{geoknife,ANY,missing,missing,missing-method}
\alias{geoknife,ANY,webdata,webprocess,missing}
\alias{geoknife,ANY,webdata,webprocess,missing-method}
\alias{geoknife,missing,ANY,missing,missing}
\alias{geoknife,missing,ANY,missing,missing-method}
\alias{geoknife,missing,missing,ANY,missing}
\alias{geoknife,missing,missing,ANY,missing-method}
\alias{geoknife,missing,missing,missing,geojob}
\alias{geoknife,missing,missing,missing,geojob-method}
\title{geoknife}
\usage{
geoknife(stencil, fabric, knife, job, ...)

\S4method{geoknife}{ANY,webdata,webprocess,missing}(stencil, fabric, knife, job,
  ...)

\S4method{geoknife}{ANY,ANY,ANY,missing}(stencil, fabric, knife, job, ...)

\S4method{geoknife}{ANY,ANY,missing,missing}(stencil, fabric, knife, job, ...)

\S4method{geoknife}{missing,missing,missing,geojob}(stencil, fabric, knife, job,
  ...)

\S4method{geoknife}{ANY,missing,missing,missing}(stencil, fabric, knife, job,
  ...)

\S4method{geoknife}{missing,ANY,missing,missing}(stencil, fabric, knife, job,
  ...)

\S4method{geoknife}{missing,missing,ANY,missing}(stencil, fabric, knife, job,
  ...)
}
\arguments{
\item{stencil}{a geometric feature (the "cookie-cutter"). Many input types are supported. [list them]}

\item{fabric}{a dataset. Can be \code{\link{webdata}} or recognized shortname (e.g., 'prism')}

\item{knife}{(optional) a \code{\link{webprocess}} object}

\item{job}{(optional) a \code{\link{geojob}} object}

\item{...}{additional arguments passed on to process.}
}
\description{
geoknife
}
\examples{
wp <- webprocess(url = 'http://cida-test.er.usgs.gov/gdp/process/WebProcessingService')
wd <- webdata('prism')
wg <- simplegeom(c(-89,42))
geoknife(stencil = c(-89,42), fabric = 'prism')
sp <- quick_sp()
job <- geoknife(stencil = sp, fabric = wd, knife = wp)
check(job)
}

