% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geoknife}
\alias{geoknife}
\title{Create geoknife object}
\usage{
geoknife()
}
\description{
Create geoknife object
}
\examples{
\dontrun{
gk <- geoknife() # create geoknife object
gk # print geoknife object

linearRing = bufferPoint(c(-111.48,36.95))
setFeature(gk) <-list(LinearRing=linearRing)

#get a list of available processing algorithms
getAlgorithms(gk)

#set processing algorithm to feature weighted grid statistics (unweighted will likely fail, because the ring won't intersect the centroids)
setAlgorithm(gk) <- getAlgorithms(gk)[4] # feature weighted

# set the post inputs for the processing dataset
setProcessInputs(gk) <- list('DATASET_ID'='Downward_longwave_radiation_flux_surface',
                                       'DATASET_URI'='dods://igsarm-cida-thredds1.er.usgs.gov:8081/qa/thredds/dodsC/nldas/best',
                                       'TIME_START'='2010-01-01T00:00:00Z',
                                       'TIME_END'='2010-01-01T23:00:00Z',
                                       'DELIMITER'='TAB')
gk # print geoknife object contents

# kick off your request
gk <- startProcess(gk)
status.gk  <-  checkProcess(gk)
checkProcess(gk) # check again and print to screen
}
}
\keyword{geoknife}

