% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/01-geoknife.R
\name{geoknife}
\alias{geoknife}
\title{Create geoknife object}
\usage{
geoknife()
}
\description{
Create geoknife object
}
\examples{
\dontrun{
gk <- geoknife() # create geoknife object
gk # print geoknife object

linearRing = bufferPoint(c(-111.48,36.95))
setFeature(gk) <-list(LinearRing=linearRing)

#get a list of available processing algorithms
getAlgorithms(gk)

#set processing algorithm to feature weighted grid statistics (unweighted will likely fail, because the ring won't intersect the centroids)
setAlgorithm(gk) <- getAlgorithms(gk)[4] # feature weighted

# set the post inputs for the processing dataset
setProcessInputs(gk) <- list('DATASET_ID'='prcp',
                                       'DATASET_URI'='http://thredds.daac.ornl.gov/thredds/dodsC/daymet-agg/daymet-agg.ncml',
                                       'TIME_START'='2010-01-01T00:00:00Z',
                                       'TIME_END'='2010-01-03T00:00:00Z',
                                       'DELIMITER'='TAB')
gk # print geoknife object contents

# kick off your request
gk <- startProcess(gk)
status.gk  <-  checkProcess(gk)
checkProcess(gk) # check again and print to screen
}
}
\author{
Jordan S. Read
}
\keyword{geoknife}

