% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom-webgeom.R
\docType{methods}
\name{geom<-}
\alias{geom}
\alias{geom<-}
\alias{geom<-,webgeom-method}
\title{the geom of an object}
\usage{
geom(.Object) <- value

geom(.Object)

\S4method{geom}{webgeom}(.Object) <- value
}
\arguments{
\item{.Object}{a \linkS4class{webgeom} object}

\item{value}{a geom}
}
\description{
the geom of an object
}

