% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getAlgorithms}
\alias{getAlgorithms}
\title{Get processing algorithms}
\usage{
getAlgorithms(.Object)
}
\arguments{
\item{.Object}{a \code{geoknife} object with a valid WPS url.}
}
\value{
list of available algorithms for the \code{geoknife} WPS url.
}
\description{
Get algorithms for \code{geoknife}
}
\details{
getAlgorithms is a method for finding algorithm names and locations for
a \code{geoknife} object.
}
\examples{
gk <- geoknife() # create geoknife object
getAlgorithms(gk) # list Geo Data Portal processing algorithms
}
\author{
Jordan S. Read
}
\seealso{
\code{setAlgorithm<-}
}
\keyword{getAlgorithms}

