\docType{methods}
\name{getAlgorithms}
\alias{getAlgorithms}
\title{getAlgorithms}
\arguments{
  \item{\code{rGDP}}{object with a valid WPS url.}
}
\value{
  list of available algorithms for the \code{rGDP} WPS url.
}
\description{
  a \code{rGDP} method for finding algorithm names and
  locations from a valid WPS endpoint.
}
\keyword{getAlgorithms}

