\docType{methods}
\name{getAlgorithms}
\alias{getAlgorithms}
\title{Get processing algorithms}
\arguments{
  \item{\code{rGDP}}{object with a valid WPS url.}
}
\value{
  list of available algorithms for the \code{rGDP} WPS url.
}
\description{
  Get algorithms for \code{rGDP}
}
\details{
  getAlgorithms is a method for finding algorithm names and
  locations for a \code{rGDP} object.
}
\seealso{
  \code{setAlgorithm}
}
\keyword{getAlgorithms}

