\docType{methods}
\name{getAttributes}
\alias{getAttributes}
\title{get attributes from a shapefile at a web location}
\arguments{
  \item{\code{rGDP}}{object with a valid WFS url.}

  \item{a}{valid shapefile name.}
}
\value{
  list of attributes for the given shapefile at the
  \code{rGDP} WFS url.
}
\description{
  get attributes from a shapefile at a web location
}
\details{
  a \code{rGDP} method for finding attribute names for a
  given shapefile at a valid WFS endpoint.
}
\keyword{getAttributes}

