\docType{methods}
\name{getAttributes}
\alias{getAttributes}
\title{getAttributes}
\arguments{
  \item{\code{rGDP}}{object with a valid WFS url.}

  \item{a}{valid shapefile name.}
}
\value{
  list of attributes for the given shapefile at the
  \code{rGDP} WFS url.
}
\description{
  function for rGDP
}
\keyword{getAttributes}

