\docType{methods}
\name{getDataIDs}
\alias{getDataIDs}
\title{getDataIDs}
\arguments{
  \item{\code{rGDP}}{object with a valid dataset URI.}

  \item{cachedResponse}{an optional input to allow cached
  response. Default as FALSE}
}
\value{
  list of dataset IDs for the \code{rGDP} dataset URI.
}
\description{
  a \code{rGDP} method for finding dataset IDs (values) for
  an rGDP object with a valid URI.
}
\keyword{getDataIDs}

