\docType{methods}
\name{getDataIDs}
\alias{getDataIDs}
\title{Find variables from dataset}
\arguments{
  \item{\code{rGDP}}{object with a valid dataset URI.}

  \item{cachedResponse}{an optional input to allow cached
  response. Default as FALSE}
}
\value{
  list of dataset IDs for the \code{rGDP} dataset URI.
}
\description{
  Find variables from \code{rGDP} dataset
}
\details{
  getDataIDs is a method for finding dataset variables from
  a \code{rGDP} object with a valid dataset URI. Use
  setPostInputs to set a dataset URI.
}
\seealso{
  \code{setPostInputs}
}
\keyword{getDataIDs}

