% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getDataIDs}
\alias{getDataIDs}
\title{Find variables from dataset}
\usage{
getDataIDs(.Object, cachedResponse)
}
\arguments{
\item{.Object}{a \code{geoknife} object with a valid dataset URI.}

\item{cachedResponse}{an optional input to allow cached response. Default as FALSE}
}
\value{
list of dataset IDs for the \code{geoknife} dataset URI.
}
\description{
Find variables from \code{geoknife} dataset
}
\details{
getDataIDs is a method for finding dataset variables from a
\code{geoknife} object with a valid dataset URI. Use setProcessInputs to set a dataset URI.
}
\examples{
gk <- geoknife() # create geoknife object
algorithm <- list("Area Grid Statistics (weighted)"=
"gov.usgs.cida.gdp.wps.algorithm.FeatureWeightedGridStatisticsAlgorithm")
setAlgorithm(gk) <- algorithm
setProcessInputs(gk) <- list("DATASET_URI" = "dods://igsarm-cida-thredds1.er.usgs.gov:8081/qa/thredds/dodsC/nldas/best")
getDataIDs(gk)
}
\author{
Jordan S. Read
}
\seealso{
\code{setProcessInputs}
}
\keyword{methods}

