\docType{methods}
\name{getValues}
\alias{getValues}
\title{getValues}
\arguments{
  \item{\code{rGDP}}{object with a valid WFS url.}

  \item{a}{valid shapefile name.}

  \item{a}{valid attribute name for the shapefile.}
}
\value{
  list of values for the given shapefile attribute at the
  \code{rGDP} WFS url.
}
\description{
  function for rGDP
}
\keyword{getValues}

