\docType{methods}
\name{getValues}
\alias{getValues}
\title{get values from a shapefile at a web location}
\arguments{
  \item{\code{rGDP}}{object with a valid WFS url.}

  \item{a}{valid shapefile name.}

  \item{a}{valid attribute name for the shapefile.}
}
\value{
  list of values for the given shapefile attribute at the
  \code{rGDP} WFS url.
}
\description{
  get values from a shapefile at a web location
}
\details{
  a \code{rGDP} method for finding value names for a given
  shapefile at a valid WFS endpoint.
}
\keyword{getValues}

