% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getValues.R
\docType{methods}
\name{getValues}
\alias{getValues}
\title{get values from a shapefile at a web location}
\usage{
getValues(.Object, shapefile, attribute)
}
\arguments{
\item{.Object}{a \code{geoknife} object with a valid WFS url.}

\item{shapefile}{a valid shapefile name.}

\item{attribute}{a valid attribute name for the shapefile.}
}
\value{
list of values for the given shapefile attribute at the \code{geoknife} WFS url.
}
\description{
get values from a shapefile at a web location
}
\details{
a \code{geoknife} method for finding value names for a given shapefile at a valid WFS endpoint.
}
\examples{
gk<- geoknife() # create geoknife object
shps <- getShapefiles(gk) # get shapefile names
atts <- getAttributes(gk,shapefile=shps[2])
getValues(gk,shapefile=shps[2],attribute=atts[4])
}
\author{
Jordan S. Read
}
\keyword{methods}

