% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inputs-webprocess.R
\docType{methods}
\name{inputs<-}
\alias{inputs}
\alias{inputs,webprocess-method}
\alias{inputs<-}
\alias{inputs<-,webprocess-method}
\title{inputs of webprocess}
\usage{
inputs(.Object, ...) <- value

inputs(.Object, ...)

\S4method{inputs}{webprocess}(.Object, ...)

\S4method{inputs}{webprocess}(.Object, ...) <- value
}
\arguments{
\item{.Object}{a webprocessing object}

\item{...}{arguments matching fields in .Object's processInputs slot}
}
\description{
access or set the inputs of a webprocess
}
\examples{
wp <- webprocess(DELIMITER = 'TAB')
inputs(wp)

inputs(wp) <- list(DELIMITER = 'COMMA', SUMMARIZE_FEATURE_ATTRIBUTE = 'false')
inputs(wp)
inputs(wp, "DELIMITER")
}
\keyword{internal}

