% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadOutput.R
\docType{methods}
\name{loadOutput}
\alias{loadOutput}
\alias{loadOutput,geojob-method}
\title{parse process output into R environment}
\usage{
loadOutput(.Object, ...)

\S4method{loadOutput}{geojob}(.Object, ...)
}
\arguments{
\item{.Object}{a \code{\link{geojob}} object with a successful processID.
(See \code{\link{check}}).}

\item{...}{additional arguments passed to parsers (e.g., keep.units = TRUE)}
}
\value{
data.frame of timeseries values.
}
\description{
parse process output into R environment
}
\details{
a \code{geojob} method for loading data into R from a completed processing request
}
\examples{
\dontrun{
job <- geoknife(stencil = c(-89,42), fabric = 'prism', wait = TRUE)
loadOutput(job, with.units = TRUE) # load and print output
}
}
\author{
Jordan S. Read
}
\keyword{methods}

