% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTimeseries.R
\name{parseTimeseries}
\alias{parseTimeseries}
\title{parse timeseries file into R environment}
\usage{
parseTimeseries(file, delim, with.units = FALSE)
}
\arguments{
\item{file}{a \linkS4class{geojob} timeseries processing result file location
(See \code{\link{download}}).}

\item{delim}{the file delimiter}

\item{with.units}{boolean for including a units column in returned data.frame (default = \code{FALSE})}
}
\value{
a data.frame of timeseries values.
}
\description{
a function for loading data into R from a file (or URL) from a completed processing request
}
\examples{
local_file <- system.file('extdata','tsv_linear_ring.tsv', package = 'geoknife')
output <- parseTimeseries(local_file, delim = '\t')
}
\seealso{
\code{\link{check}}, \code{\link{download}}, \code{\link{parseCategorical}}
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}
