% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseTimeseries.R
\name{parseTimeseries}
\alias{parseTimeseries}
\title{parse timeseries file into R environment}
\usage{
parseTimeseries(file, delim)
}
\arguments{
\item{file}{a \code{\link{geoknife}} timeseries processing result file location
(See \code{\link{checkProcess}}).}

\item{delim}{the file delimiter}
}
\value{
list of timeseries values.
}
\description{
parse timeseries file into R environment
}
\details{
a function for loading data into R from a file (or URL) from a completed processing request
}
\examples{
local_file <- system.file('extdata','tsv_linear_ring.tsv', package = 'geoknife')
output <- parseTimeseries(local_file, delim = '\\t')
}
\author{
Luke A. Winslow, Jordan S. Read
}
\keyword{methods}

