% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query-webgeom.R
\name{query}
\alias{query}
\title{query webgeom for various fields}
\usage{
query(.Object, field, ...)
}
\arguments{
\item{.Object}{a \code{webgeom} object.}

\item{field}{a plural parameter name in webgeom (e.g., 'geoms')}

\item{...}{additional arguments passed to GET}
}
\value{
a character vector of values corresponding to the query field specified
}
\description{
query webgeom for various fields
}
\details{
a \code{webgeom} method for finding shapefile names at a valid WFS endpoint.
}
\examples{
\dontrun{
wg <- webgeom()
query(wg, 'geoms')
}
}
\author{
Jordan S. Read
}
\keyword{methods}

