% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoknifeUtils.R
\name{retryVERB}
\alias{retryVERB}
\title{simple retry for httr VERBs}
\usage{
retryVERB(VERB, url, ..., retries = gconfig("retries"))
}
\arguments{
\item{VERB}{\code{httr} function}

\item{url}{a url}

\item{retries}{number of times to try before failing}

\item{\dots}{additional args passed into VERB}
}
\description{
Useful for intermittent server issues, when a retry will get the job done
}
\keyword{internal}

