\docType{methods}
\name{setAlgorithm}
\alias{setAlgorithm}
\title{setAlgorithm}
\arguments{
  \item{An}{\code{rGDP} object.}

  \item{a}{list for a valid algorithm, including values for
  name & location}
}
\value{
  An \code{rGDP} object.
}
\description{
  function for rGDP
}
\keyword{setAlgorithm}

