\docType{methods}
\name{setFeature}
\alias{setFeature}
\title{Set feature geometry for processing}
\arguments{
  \item{An}{\code{geoknife} object.}

  \item{a}{list containing a valid feature collection, or a
  list of a subset of a valid feature collection.}
}
\value{
  An \code{geoknife} object.
}
\description{
  Set feature for \code{geoknife}
}
\details{
  setFeature is a method for setting or modifying the
  feature elements of for \code{geoknife} object.
}
\seealso{
  \code{setWFS}
}
\keyword{setFeature}

