\docType{methods}
\name{setProcessInputs}
\alias{setProcessInputs}
\title{set inputs for web processing}
\arguments{
  \item{An}{\code{rGDP} object.}

  \item{a}{list of valid processInputs.}
}
\value{
  An \code{rGDP} object.
}
\description{
  set inputs for web processing
}
\details{
  method for setting the (non-feature related) post inputs
  of the \code{rGDP} object.
}
\keyword{setProcessInputs}

