\docType{methods}
\name{setWFS}
\alias{setWFS}
\title{setWFS}
\arguments{
  \item{\code{rGDP}}{object.}

  \item{a}{Web Feature Service (WFS) endpoint.}
}
\value{
  An \code{rGDP} object.
}
\description{
  method for setting the web feature service (WFS) endpoint
  for a \code{rGDP} object.
}
\keyword{setWFS}

