% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-simplegeom-obj.R
\docType{methods}
\name{simplegeom}
\alias{simplegeom}
\alias{simplegeom,ANY-method}
\alias{simplegeom,missing-method}
\title{Create simplegeom object}
\usage{
simplegeom(.Object, ...)

\S4method{simplegeom}{missing}(.Object, ...)

\S4method{simplegeom}{ANY}(.Object, ...)
}
\arguments{
\item{.Object}{any object that can be coerced into \linkS4class{simplegeom}}

\item{...}{additional arguments passed to SpatialPolygonsDataFrame}
}
\value{
the simplegeom object
}
\description{
A simple geom is a simple set of geometries specified locally. See 
\code{\link{webgeom}} for web features.
}
\examples{
simplegeom(c(-88.6, 45.2))
\dontrun{
simplegeom(Srl, proj4string = CRS("+proj=longlat +datum=WGS84"))
}
simplegeom(data.frame('point1'=c(-89, 46), 'point2'=c(-88.6, 45.2)))
}
\author{
Jordan S Read
}

