% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/06-simplegeom-obj.R
\docType{methods}
\name{simplegeom}
\alias{simplegeom}
\alias{simplegeom,ANY-method}
\alias{simplegeom,simplegeom-method}
\title{create simplegeom object}
\usage{
simplegeom(...)

\S4method{simplegeom}{ANY}(...)
}
\value{
the simplegeom object simple feature
}
\description{
A class representing a simple feature.
}
\section{Slots}{

\describe{
\item{\code{polys}}{value of type}

\item{\code{IDs}}{value of type (same length as polys)}

\item{\code{proj}}{value of type}
}}
\author{
Jordan S Read
}

