% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/06-simplegeom-obj.R
\docType{methods}
\name{simplegeom}
\alias{simplegeom}
\alias{simplegeom,ANY-method}
\alias{simplegeom,simplegeom-method}
\title{create simplegeom object}
\usage{
simplegeom(...)

\S4method{simplegeom}{ANY}(...)
}
\arguments{
\item{...}{additional arguments passed to SpatialPolygonsDataFrame}

\item{value}{a numeric vector (for long,lat point) or data.framed with long, lat pairs, [future support for more]}
}
\value{
the simplegeom object
}
\description{
A class representing a simple feature.
}
\section{Slots}{

\describe{
\item{\code{DRAW_NAMESPACE}}{location of polygrom draw namespace}

\item{\code{DRAW_SCHEMA}}{url for draw schema}

\item{\code{sp}}{object of class \code{\link[sp]{SpatialPolygons}}}
}}
\examples{
as(c(-88.6, 45.2), "simplegeom")
\dontrun{
simplegeom(Srl, proj4string = CRS("+proj=longlat +datum=WGS84"))
}
as(data.frame('point1'=c(-89, 46), 'point2'=c(-88.6, 45.2)), "simplegeom")
}
\author{
Jordan S Read
}

