% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/start-geojob.R
\docType{methods}
\name{start}
\alias{start}
\alias{start,geojob-method}
\title{start a geo-web processing request}
\usage{
start(.Object)

\S4method{start}{geojob}(.Object)
}
\arguments{
\item{.Object}{a \linkS4class{geojob} object}
}
\value{
A \linkS4class{geojob} object with an active GDP process request.
}
\description{
Start process for \code{geojob}
}
\details{
start is a method for submitting a geo-web processing request.
}
\examples{
wd <- webdata('prism')
wg <- quick_wg()
gj <- geojob()
xml(gj) <- XML(wg, wd, webprocess())
url(gj) <- url(wp)
gj <- start(gj)
}
\seealso{
\code{check}
}
\keyword{methods}

