\docType{methods}
\name{startProcess}
\alias{startProcess}
\title{Submit a GDP web processing request}
\arguments{
  \item{\code{rGDP}}{object ot be used to formulate GDP
  process request.}
}
\value{
  An \code{rGDP} object with an active GDP process request.
}
\description{
  Start process for \code{rGDP}
}
\details{
  startProcess is a method for submitting a web processing
  request for the \code{rGDP} object.
}
\seealso{
  \code{checkProcess}
}
\keyword{startProcess}

