% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/url-generic.R
\docType{methods}
\name{url<-}
\alias{url}
\alias{url,geojob-method}
\alias{url,webdata-method}
\alias{url,webgeom-method}
\alias{url,webprocess-method}
\alias{url<-}
\alias{url<-,geojob-method}
\alias{url<-,webdata-method}
\alias{url<-,webgeom-method}
\alias{url<-,webprocess-method}
\title{the url of an object}
\usage{
url(.Object)
url(.Object)<- value

url(.Object)

\S4method{url}{webdata}(.Object) <- value

\S4method{url}{webgeom}(.Object) <- value

\S4method{url}{webdata}(.Object)

\S4method{url}{webgeom}(.Object)

\S4method{url}{geojob}(.Object) <- value

\S4method{url}{geojob}(.Object)

\S4method{url}{webprocess}(.Object) <- value

\S4method{url}{webprocess}(.Object)
}
\arguments{
\item{.Object}{a webgeom, webdata, geojob, or webprocess object}

\item{value}{a url}
}
\description{
the url of an object
}

