% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/url-generic.R
\docType{methods}
\name{url<-}
\alias{url}
\alias{url,ANY-method}
\alias{url,datagroup-method}
\alias{url<-}
\alias{url<-,ANY-method}
\alias{url<-,webprocess-method}
\title{the url of an object}
\usage{
url(.Object)
url(.Object)<- value

url(.Object)

\S4method{url}{ANY}(.Object) <- value

\S4method{url}{webprocess}(.Object) <- value

\S4method{url}{datagroup}(.Object)

\S4method{url}{ANY}(.Object)
}
\arguments{
\item{.Object}{a \linkS4class{webgeom}, \linkS4class{webdata},
\linkS4class{geojob}, or \linkS4class{webprocess} object}

\item{value}{a url}
}
\description{
the url of an object
}

