% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-generic.R
\name{url<-}
\alias{url<-}
\alias{url}
\alias{url<-,ANY-method}
\alias{url<-,webprocess-method}
\alias{url,character-method}
\alias{url,missing-method}
\alias{url,datagroup-method}
\alias{url,ANY-method}
\title{the url of an object}
\usage{
url(.Object) <- value

url(.Object, ...)

\S4method{url}{ANY}(.Object) <- value

\S4method{url}{webprocess}(.Object) <- value

\S4method{url}{character}(.Object, ...)

\S4method{url}{missing}(.Object, ...)

\S4method{url}{datagroup}(.Object, ...)

\S4method{url}{ANY}(.Object, ...)
}
\arguments{
\item{.Object}{a \linkS4class{webgeom}, \linkS4class{webdata},}

\item{value}{a url}

\item{\dots}{additional arguments that would be passed to the masked base::url function.
These are only used when the .Object argument is character or missing
\linkS4class{geojob}, or \linkS4class{webprocess} object}
}
\description{
get or set the url of an object
}
