% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variables-webdata.R
\docType{methods}
\name{variables}
\alias{variables}
\alias{variables,webdata-method}
\alias{variables<-}
\alias{variables<-,webdata,character-method}
\title{the variables of a webdata object}
\usage{
variables(.Object)
variables(.Object) <- value

variables(.Object) <- value

\S4method{variables}{webdata}(.Object)

\S4method{variables}{webdata,character}(.Object) <- value
}
\arguments{
\item{.Object}{a \code{\link{webdata}} object}

\item{value}{a character vector for variables}
}
\description{
the variables of a webdata object
}

