% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/01-webdata-obj.R
\docType{class}
\name{webdata-class}
\alias{webdata-class}
\title{webdata class}
\description{
webdata class
}
\section{Slots}{

\describe{
\item{\code{times}}{vector of POSIXct dates (specifying start and end time of processing)}

\item{\code{url}}{URL of web data}

\item{\code{variables}}{variable(s) used for processin from dataset}
}}

