% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/01-webdata-obj.R, R/02-webgeom-obj.R
\docType{methods}
\name{webdata}
\alias{times}
\alias{times,webdata-method}
\alias{times<-}
\alias{times<-,webdata-method}
\alias{url}
\alias{url,webdata-method}
\alias{variables}
\alias{variables,webdata-method}
\alias{variables<-}
\alias{variables<-,webdata-method}
\alias{webdata}
\alias{webdata,ANY-method}
\alias{webdata,webdata-method}
\title{create webdata object}
\usage{
webdata(...)

\S4method{webdata}{ANY}(...)

times(.Object) <- value

\S4method{times}{webdata}(.Object) <- value

times(.Object)

variables(.Object)

\S4method{variables}{webdata}(.Object)

variables(.Object) <- value

\S4method{variables}{webdata}(.Object) <- value

url(.Object)
}
\arguments{
\item{...}{additional arguments passed initialize method (e.g., times vector)}

\item{.Object}{a \code{\link{webdata}} object}

\item{value}{a POSIXct vector}

\item{.Object}{a \code{\link{webdata}} object}

\item{.Object}{a \code{\link{webdata}} object}

\item{value}{a character vector for variables}
}
\value{
the webdata object representing a dataset and parameters
}
\description{
A class representing a web dataset.
}
\section{Slots}{

\describe{
\item{\code{times}}{value of type \code{"POSIXct"}, start and stop dates for data}

\item{\code{url}}{value of type \code{"character"}, the web location for the dataset}

\item{\code{variable}}{value of type \code{"character"}, the variable(s) for data}
}}
\examples{
wd <- webdata()
times(wd) <- as.POSIXct(c("2012-11-04", "2012-11-12"))
times(wd)[1] <- as.POSIXct("2012-11-04")
}
\author{
Jordan S Read
}

