% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/01-webdata-obj.R, R/02-webgeom-obj.R
\docType{methods}
\name{webdata}
\alias{times}
\alias{times,webdata-method}
\alias{times<-}
\alias{times<-,webdata-method}
\alias{url}
\alias{url,webdata-method}
\alias{url<-}
\alias{url<-,webdata-method}
\alias{variables}
\alias{variables,webdata-method}
\alias{variables<-}
\alias{variables<-,webdata-method}
\alias{webdata}
\alias{webdata,ANY-method}
\alias{webdata,webdata-method}
\title{create webdata object}
\usage{
webdata(...)

\S4method{webdata}{ANY}(...)

times(webdata()) <- as.POXIXct(c("2012-11-04", "2012-11-12"))
times(webdata())[1] <- as.POXIXct("2012-11-04")

times(.Object)

variables(.Object)

variables(.Object) <- value

url(.Object, values) <- value

url(.Object)
}
\value{
the webdata object representing a dataset and parameters
}
\description{
A class representing a web dataset.
}
\section{Slots}{

\describe{
\item{\code{times}}{value of type \code{"POSIXct"}, start and stop dates for data}

\item{\code{url}}{value of type \code{"character"}, the web location for the dataset}

\item{\code{variable}}{value of type \code{"character"}, the variable(s) for data}
}}
\author{
Jordan S Read
}

