% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/02-webgeom-obj.R
\docType{class}
\name{webgeom-class}
\alias{webgeom-class}
\title{webgeom class}
\description{
webgeom class
}
\section{Slots}{

\describe{
\item{\code{url}}{URL of web feature service endpoint}

\item{\code{geom}}{character for geometric feature name}

\item{\code{attribute}}{character for feature attribute (used for filtering and naming in output)}

\item{\code{IDs}}{attribute IDs to be used in processing (a subset, or all if NA)}

\item{\code{wfs_version}}{the web feature service version to use.}

\item{\code{id}}{job identifier}
}}

