% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/02-webgeom-obj.R
\docType{methods}
\name{webgeom}
\alias{url<-}
\alias{url<-,webgeom-method}
\alias{webgeom}
\alias{webgeom,ANY-method}
\alias{webgeom,webgeom-method}
\title{create webgeom object}
\usage{
webgeom(...)

\S4method{webgeom}{ANY}(...)

url(.Object, values) <- value

\S4method{url}{webgeom}(.Object, values) <- value
}
\value{
the webgeom object representing a dataset and parameters
}
\description{
A class representing a web dataset.
}
\section{Slots}{

\describe{
\item{\code{url}}{value of type \code{"character"}, the web location for the dataset}

\item{\code{variable}}{value of type \code{"character"}, the variable(s) for data}
}}
\author{
Jordan S Read
}

