% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/04-webprocess-obj.R
\docType{class}
\name{webprocess-class}
\alias{webprocess-class}
\title{webprocess class}
\description{
webprocess class
}
\section{Slots}{

\describe{
\item{\code{url}}{URL for webprocessing service}

\item{\code{algorithm}}{a list for algorithm used}
}}

