% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/04-webprocess-obj.R
\docType{methods}
\name{webprocess}
\alias{webprocess}
\alias{webprocess,ANY-method}
\title{create webprocess object}
\usage{
webprocess(...)

\S4method{webprocess}{ANY}(...)
}
\arguments{
\item{...}{additional arguments passed initialize method (e.g., \code{url})}
}
\value{
the webprocess object
}
\description{
create webprocess object
}
\author{
Jordan S Read
}

