% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04-webprocess-obj.R
\docType{methods}
\name{webprocess}
\alias{webprocess}
\alias{webprocess,ANY-method}
\alias{webprocess,missing-method}
\title{create webprocess object}
\usage{
webprocess(.Object, ...)

\S4method{webprocess}{missing}(.Object, ...)

\S4method{webprocess}{ANY}(.Object, ...)
}
\arguments{
\item{.Object}{any object that can be coerced into \linkS4class{webdprocess}}

\item{...}{additional arguments passed initialize method (e.g., \code{url}, \code{version})}
}
\value{
the webprocess object
}
\description{
create webprocess object
}
\author{
Jordan S Read
}

