% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processInputsToXML.R
\docType{methods}
\name{XML}
\alias{XML}
\alias{XML,ANY,webdata,webprocess-method}
\alias{XML,webgeom-method}
\title{XML from set of objects}
\usage{
XML(stencil, fabric, knife)

\S4method{XML}{ANY,webdata,webprocess}(stencil, fabric, knife)
}
\arguments{
\item{stencil}{a \code{\link{webdata}} OR \code{\link{simplegeom}} object}

\item{fabric}{a \code{\link{webdata}} object}

\item{knife}{a \code{\link{webprocess}} object}
}
\value{
XML as ?string?
}
\description{
XML from set of objects
}
\examples{
wd <- webdata('prism',times = as.POSIXct(c('2001-01-01','2002-02-05')))
wg <- webgeom('state::Wisconsin')
\dontrun{
XML(wg, wd, webprocess())
sg <- simplegeom(c(-89,45))
XML(sg, wd, webprocess())
}
}

